requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic==0.84.0
mediapipe==0.10.14
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.8.0
PyJWT[crypto]==2.9.0
cryptography==43.0.3
httpx==0.27.2
boto3==1.35.0
supabase>=2.0.0
upstash-redis>=1.0.0
sentry-sdk[fastapi]>=2.0.0
